requirements.txt
streamlit>=1.35.0
langgraph>=0.4.0
langchain-core>=0.3.0
tavily-python>=0.5.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
google-genai>=1.0.0
groq>=0.13.0

# Dev / testing
pytest>=8.0.0
